+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Streaming with cache enabled.
+
 ### Changed
 - Use OpenAI response_format instead of separators in the prompt.
 - Switch to cohere reranker v3 and `retriever_k = 500`.